# docs/swagger.yaml
basePath: /api-market-testnet
definitions:
...
  models.ConfirmationState:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - ConfirmationStateUnconfirmed
    - ConfirmationStateConfirmed
...
  request.MintPinInfo:
    properties:
      address:
        type: string
      outRaw:
        type: string
      pinId:
        type: string
      pinUtxoIndex:
        type: integer
      pinUtxoOutValue:
        type: integer
      pinUtxoTxId:
        type: string
      pkScript:
        type: string
    type: object
  request.Mrc20DeployCommitRequest:
    properties:
      commitTxOutIndex:
        type: integer
      commitTxRaw:
        type: string
      orderId:
        type: string
      revealTxRaw:
        description: RevealPrePsbtRaw string `json:"revealPrePsbtRaw"`
        type: string
    type: object
  request.Mrc20DeployPreRequest:
    properties:
      address:
        type: string
      networkFeeRate:
        type: integer
      payload:
        type: string
    type: object
...
  request.Mrc20MintCommitRequest:
    properties:
      commitTxOutIndex:
        type: integer
      commitTxRaw:
        type: string
      orderId:
        type: string
      revealPrePsbtRaw:
        type: string
    type: object
  request.Mrc20MintPreRequest:
    properties:
      mintPins:
        items:
          $ref: '#/definitions/request.MintPinInfo'
        type: array
      networkFeeRate:
        type: integer
      outAddress:
        type: string
      outValue:
        type: integer
      tickerId:
        type: string
    type: object
  request.Mrc20OutInfo:
    properties:
      address:
        type: string
      amount:
        type: string
      outValue:
        type: integer
      pkScript:
        type: string
    type: object
  request.Mrc20TransferCommitRequest:
    properties:
      commitTxOutIndex:
        type: integer
      commitTxRaw:
        type: string
      orderId:
        type: string
      revealPrePsbtRaw:
        type: string
    type: object
  request.Mrc20TransferPreRequest:
    properties:
      changeAddress:
        type: string
      changeOutValue:
        type: integer
      mrc20Outs:
        items:
          $ref: '#/definitions/request.Mrc20OutInfo'
        type: array
      networkFeeRate:
        type: integer
      tickerId:
        type: string
      transfers:
        items:
          $ref: '#/definitions/request.TransferMrc20Utxo'
        type: array
    type: object
...
  request.TransferMrc20Utxo:
    properties:
      address:
        type: string
      amount:
        type: string
      outRaw:
        type: string
      pkScript:
        type: string
      tickerId:
        type: string
      utxoIndex:
        type: integer
      utxoOutValue:
        type: integer
      utxoTxId:
        type: string
    type: object
...
  respond.FetchMrc20OpOrdersResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.OpOrderInfoResp'
        type: array
      total:
        type: integer
    type: object
...
  respond.Mrc20DeployCommitResp:
    properties:
      commitTxId:
        type: string
      orderId:
        type: string
      revealTxId:
        type: string
    type: object
  respond.Mrc20DeployPreResp:
    properties:
      extra:
        description: |-
          RevealPrePsbtRaw string      `json:"revealPrePsbtRaw"`
          RevealInputIndex int64       `json:"revealInputIndex"`
      minerFee:
        type: integer
      orderId:
        type: string
      revealAddress:
        type: string
      serviceFee:
        type: integer
      totalFee:
        type: integer
    type: object
...
  respond.Mrc20MintCommitResp:
    properties:
      commitTxId:
        type: string
      orderId:
        type: string
      revealTxId:
        type: string
    type: object
  respond.Mrc20MintPreResp:
    properties:
      extra: {}
      orderId:
        type: string
      revealAddress:
        type: string
      revealFee:
        type: integer
      revealInputIndex:
        type: integer
      revealPrePsbtRaw:
        type: string
      serviceAddress:
        type: string
      serviceFee:
        type: integer
      totalFee:
        type: integer
    type: object
...
  respond.Mrc20TransferCommitResp:
    properties:
      commitTxId:
        type: string
      orderId:
        type: string
      revealTxId:
        type: string
    type: object
  respond.Mrc20TransferPreResp:
    properties:
      extra: {}
      orderId:
        type: string
      revealAddress:
        type: string
      revealFee:
        type: integer
      revealInputIndex:
        type: integer
      revealPrePsbtRaw:
        type: string
      serviceAddress:
        type: string
      serviceFee:
        type: integer
      totalFee:
        type: integer
    type: object
...
  respond.OpOrderInfoResp:
    properties:
      amtPerMint:
        type: string
      blockHeight:
        type: integer
      confirmationState:
        $ref: '#/definitions/models.ConfirmationState'
      decimals:
        type: string
      deployState:
        description: 0-pending, 1-success, 2-fail
        type: integer
      deployerAddress:
        type: string
      deployerMetaId:
        type: string
      deployerUserInfo:
        $ref: '#/definitions/respond.UserInfo'
      metaData:
        type: string
      mintCount:
        type: string
      mintState:
        description: 0-pending, 1-success, 2-fail
        type: integer
      opOrderType:
        type: string
      orderId:
        type: string
      payCheck:
        type: string
      pinCheck:
        type: string
      premineCount:
        type: string
      qual:
        type: string
      startBlockHeight:
        type: string
      tick:
        type: string
      tickId:
        type: string
      tickName:
        type: string
      timestamp:
        type: integer
      totalMinted:
        type: string
      txId:
        type: string
      usedPins:
        description: |-
          QualPath          string                   `json:"qualPath"`
          QualLevel         int64                    `json:"qualLevel"`
          QualCount         int64                    `json:"qualCount"`
        items:
          type: string
        type: array
    type: object
...
  respond.UserInfo:
    properties:
      avatar:
        type: string
      name:
        type: string
    type: object
info:
  contact:
    name: API Support
  description: MetaID-Market API Service
  title: MetaID-Market API Service
  version: "1.0"
paths:
...
  /api/v1/inscribe/mrc20/deploy/commit:
    post:
      description: Inscribe Mrc20 deploy commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20DeployCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20DeployCommitResp'
      summary: Inscribe Mrc20 deploy commit
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/deploy/pre:
    post:
      description: Inscribe Mrc20 deploy pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20DeployPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20DeployPreResp'
      summary: Inscribe Mrc20 deploy pre
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/mint/commit:
    post:
      description: Inscribe Mrc20 Mint Commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20MintCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20MintCommitResp'
      summary: Inscribe Mrc20 Mint Commit
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/mint/pre:
    post:
      description: Inscribe Mrc20 Mint Pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20MintPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20MintPreResp'
      summary: Inscribe Mrc20 Mint Pre
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/orders:
    get:
      description: Fetch inscribe mrc20 orders
      parameters:
      - description: default:deploy, mint, transfer
        in: query
        name: opOrderType
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: default:0
        in: query
        name: cursor
        type: integer
      - description: default:10, max:50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.FetchMrc20OpOrdersResp'
      summary: Fetch inscribe mrc20 orders
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/transfer/commit:
    post:
      description: Inscribe Mrc20 Transfer Commit
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20TransferCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TransferCommitResp'
      summary: Inscribe Mrc20 Transfer Commit
      tags:
      - Inscribe-Mrc20
  /api/v1/inscribe/mrc20/transfer/pre:
    post:
      description: Inscribe Mrc20 Transfer Pre
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.Mrc20TransferPreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TransferPreResp'
      summary: Inscribe Mrc20 Transfer Pre
      tags:
      - Inscribe-Mrc20
...
schemes:
- https
swagger: "2.0"